+++ src/components/index.js
@@ -2,7 +2,7 @@ export { default as PageLayout } from './layout';
 export { default as AdminNavbar } from './navbar';
 export { default as AdminSidebar } from './sidebar';
 export { default as OrdersDashboard } from './orders';
-export { default as ProductsDashboard } from './products';
+export { default as Products } from './products';
 export { default as ViewOrders } from './orders/view-orders';
 export {default as RolesAndPermissions} from './roles-and-permissions';
 export {default as Welcome} from './dashboard/welcome'
